Alert Monitor Alert - Move Rest Client to Common
