+++ Tasks/AzureMonitorAlertsV0/make.json
@@ -1,18 +1,10 @@
 {
     "common": [
         {
-            "module": "../Common/azurestack-common",
+            "module": "../Common/azure-arm-rest",
             "type": "node",
             "dest": "./",
             "compile" : true
-        }
-    ],
-    "rm": [
-        {
-            "items": [
-                "node_modules/azurestack-common/node_modules/vsts-task-lib"
-            ],
-            "options": "-Rf"
         }
     ]
 }
