+++ Tasks/AzureMonitorAlertsV0/task.json
@@ -13,7 +13,7 @@
 	"version": {
 		"Major": 0,
 		"Minor": 1,
-		"Patch": 2
+		"Patch": 3
 	},
 	"minimumAgentVersion": "2.111.0",
 	"instanceNameFormat": "Configure Azure Alerts : $(ResourceName)",
